Migrate from abandoned java-gitlab-api to gitlab4j-api (#659)

+++ src/main/kotlin/svnserver/ext/gitlab/mapping/GitLabHookEvent.kt
@@ -30,7 +30,7 @@ class GitLabHookEvent {
     val pathWithNamespace: String? = null
 
     @JsonProperty("project_id")
-    val projectId: Int? = null
+    val projectId: Long? = null
 
     companion object {
         private val mapper = ObjectMapper()
